Add sendByDefault option for brands
When transmitting patterns for all brands with the additional patterns
option disabled, only the first pattern of the TV-B-Gone used to be
transmitted. Since this is not exactly useful, this commit introduces a
sendByDefault option for brands so that we can avoid transmitting certain
brands if additional patterns are disabled.